Improved deck picker screen
Redesigned deck picker that gets rid of the filesystem browser. The path
to the folder in which the decks are stored are set in the preferences.
The picker screen then shows a list of all the decks.

Deck.java contains a new class to represent the top level object in the
database, the deck. It is nowhere near finished, but is supposed to
contain the functionality of the Deck class in libanki.

No new data is written to the database. Read data (due cards etc.) is
cached and might be out of date.