Progress dialog when loading a new deck
There can be quite a delay when getting the next card from a big deck.
So I added a simple progress dialog when loading a new deck, so that the
user knows something has happened when he selects a new deck. This
implied creating a new thread to do the loading while the ui thread
shows the dialog.