Asynchronous loading of card counts in deck picker
Delegates the loading of the card counts to a separate thread to improve
responsiveness when choosing the deck picker. I'm not sure how locking
for the database works so I implemented a separate locking scheme. With
this, a deck choise will not be acted upon until the counts for all the
decks have been loaded. Should be fairly easy to implement so that the
loader thread will stop loading new decks if one has already been
chosen.

Also fixed an issue about the deck not being loaded again when pressing
the "back" button, causing a force close when trying to space the next
card. It's just a quick fix in Ankidroid.onResume(), not very pretty.