Fix for issues when rotating the screen
This fixes the crash when changing screen orientation while a deck is
loading. It also saves the currently displaying card so that it doesn't
need to be reloaded if the screen changes orientation. This improves the
responsiveness.

I also moved all the card loading into a separate thread subclass in
order to make the code cleaner and to remove some redundancies.