Make sure all database cursors are properly closed
Surround all cursor statements with try...finally to make sure that the
cursor is closed at all times, even when an unexpected exception occurs.
If a cursor is not closed, java will throw an exception, but it is
silently (well, it can be seen in the log) discarded and does not cause
a force close. However, it does cause a noticeable slow-down in the execution.