Add FLAG_ACTIVITY_NEW_TASK to intent.
When starting an activity from a service, you always need to add the
FLAG_ACTIVITY_NEW_TASK, since this is not part of the same task as the
service, since the service does not have a task by itself.

This leads to an error when starting the application by clicking on the
deck name in the widget, since this actually starts the server, which in
turn starts the application.