Open the right deck when clicking on the widget.
Add the URI of the deck when invoking the StudyOptions activity.
This means that the right deck will be opens.

This currently only works if the application is not already open.
We might want to consider changing the logic of the StudyOptions
activity to always open deck if receiving the VIEW action.

Fixes issue 578: Touching the deck name in the widget should open the
corresponding deck.