Add an extra to the MAIN action to specify a deck.
This is to avoid opening the study options activity more than once. For
it not to open the same activity again but bring the existing one to the
front, the intents have to match (excluding extras).

This change allows one to specify an extra with the deck filename and
use it to start the activity with the given deck opened. Currently this
only works if the activity is not already opened.

In all other cases, the existing study option is brought to the
foreground instead. This is needed by the widget to prevent it from
opening a different copy of the study options and then making the
behavior of the back button inconsistent, i.e., it should not go back to
a previously opened deck but instead always close the app.