Correctly show the number of cards to review.
We store the due cards, which include both failed cards and those to be
reviewed. Since the widget shows the number of failed cards as well, it
should only show the number of cards to be reviewed along side it, not
the total due cards.

Issue 590: Due cards count includes failed cards count.