Strip invalid characters from deck names.
We currently sanitize the title of the deck to determine the file name
of a downloaded deck. However, this causes issues because the deck file
name must match the deck title for synchronization purposes.

The current list, based on the Anki desktop latest code, contains \, /,
and : as invalid. All the rest should just work.

Moreover, this is based on the content of a deck updated to the server
already, so it should be fine anyway (none of these characters should be
in the deck name in the first place).

Issue 586: Synchronization and handling of underscores in filenames

Issue 589: Deck with non-ASCII characters don't show up properly, and
don't sync

Issue 607: syncing problem when the deck name contains an apostrophe