Handle "Server too busy" status code.
Moreover, always handle an unknown status code as an error, since we
cannot assume the connection was successful.

Add a new error message for the "Server too busy" case.

Also, add the assumption on AnkiDroidProxy.hasDeck(String) that
connect(boolean) has been called before hand, and throw an exception if
this assumption is violated. It is not possible for the method
AnkiDroidProxy.hasDeck(String) to return a reasonable value if the call
to connect(boolean) fails, therefore, it is easier to let the caller
deal with it instead by requiring it call connect(boolean) beforehand.

Issue 598: Properly handle "AnkiWeb is too busy right now" response from
server