Add CSS classes for desktop and AnkiDroid.
This commit adds two new classes: AnkiDroid and AnkiDesktop.

The AnkiDroid is always visible in AnkiDroid and AnkiDesktop is always
invisible. This allows building a card layout which is different between
the AnkiDroid and Anki on the desktop.

Simply add:

<style>.AnkiDroid { display: none }</style>

to the card template of your deck.

Then, wrap anything that should be displayed only on the desktop with:

<div class="AnkiDesktop">
  ...
</div>

Last, wrap anything that should be displayed only on AnkiDroid with:

<div class="AnkiDroid">
  ...
</div>

The former will only be displayed when reviewing on the desktop client.
The former will only be displayed when reviewing on AnkiDroid.