Only update the widget if enabled.
There is no point in checking the status of all the decks if the widget
is not actually enabled. Moreover, this operation is quite expensive and
may slow down other operations on the database.

Keep track in a preference whether the widget is enabled and only
refresh the list of decks if the widget is actually in use.