Fix for issue 705: Sync button does not work on Honeycomb.
The underlying reason for the issue is that in Honeycomb the menu items
are generated whenever the action bar is shown instead of when the user
presses the menu key on the device; as a consequence, the StudyOptions
activity will get first a disabled Sync button in the action bar (while
the deck is loading) and this will not be enabled when the deck finally
loads (or disabled when the deck is closed).

To fix this, we need to call Activity.invalidateOptionsMenu() any time
something happens that causes the options menu options to change.

However, this method is only available is API Level 11 and above. To fix
this (and future problems of this sort). I introduced a new Compat
interface which is meant to contain functions which need to be
implemented differently on different API levels.

There are a number of implementations of this interface (currently on
CompatV3 and CompatV11) which may do different things on the different
platforms: for example, there is no need to invalidate the options menu
items on any platform before Honeycomb, so
Compat.invalidOptionsMenu(Activity) is simply an empty function in
CompatV3.