Fixes issue 760: Crash after downloading missing media.
The StyleDialog builder was incorrectly notifying the latest listener
instead of the one set at construction time, which means that if the
builder is re-used, it will actually update the listener to be used when
the dialog button is clicked.

In the case above, the "Ok" button on the missing media dialog was
actually calling the listener to the "Ok" button of the restore dialog,
causing issues because no deck was open at the time the listener is
invoked.

I removed all references to members of the builder by creating a static
nested class instead of using an anonymous inner class. This should fix
this and possibly other issues.