Issue 874: card sorting not working.
There was a mismatch between the columns and their indices in the query
used to construct the set of cards used by the card editor.

Extract that logic into a separate class and add tests to guarantee
consistency.

Also fix a null pointer exception when the SQL query fails to execute.