Issue 630: Card counts disappear after opening the options menu
We were doing something quite complex to disable the top bar when the
menu is opened. However, this causes a number of issues with Honeycomb+
since "onPrepareOptionsMenu" may be called before the menu is actually
opened.

Instead of splitting the logic into two different paths (pre- and post-
Honeycomb) I actually removed this logic.

I tested it on Donut and Gingerbread and it works fine without it.

This was introduced by:
https://github.com/flerda/Anki-Android/commit/681eda513f87074694e8df81e6dc49b21f90c41e
in order to force showing the notification area.

However, this causes flickering which is only partially fixed by the
existing implementation. Since it is still possible to switch to the
home screen to check the notifications, I think it is okay to remove it.

If we wanted a quick way to see the notification, I would suggest having
some way to toggle fullscreen from the options menu instead.