Issue 630: Card counts disappear after opening the options menu
We were doing something quite complex to disable the top bar when the
menu is opened. However, this causes a number of issues with Honeycomb+
since "onPrepareOptionsMenu" may be called before the menu is actually
opened.

Instead of splitting the logic into two different paths (pre- and post-
Honeycomb) I actually removed this logic and I tested it on Donut and
Gingerbread. It is working well on both.