Issue 937: no menu button during review on galaxy nexus
Restores the previous logic for adding the right items to the menu and
fix the underlying problem.

The problem was that we do not call "invalidateOptionsMenu()" when the
current card changes. This means that while it is initially loading, we
set the options menu to be hidden, and we never actually change it
contents.