Issue 1382: Deck description HTML
The deck description is defined to be HTML, but it is currently treated
as plain text by AnkiDroid when displaying it below the deck name.

Parse it as HTML and convert it to a spannable string with formatting.
This fixes the displaying of the string but not the editing, which still
shows the HTML and requires you to edit the HTML. Using a rich-text
editor would be better there, but beyond the scope of 2.0, so leave it
as is for now.