Always replace fragment content on deck selection.
We are creating a new StudyOptionsFragment instance and replacing
the content of the fragment on the right with it. There is no need to
check if the existing fragment is null or of the right type since we are
replacing it anyway. Also remove the check for different deck IDs since
it's possible to not have a StudyOptionsFragment loaded but still select
the same deck (e.g., when resuming activity).