Do not add font files with the wrong extension.
Since the user can put any arbitrary files in the fonts directory on the
SD card, simply ignore anything that is not a .ttf or .ttc file.

If any extraneous files are added as fonts, a corresponding @font-face
directive will be added to the stylesheet, which will actually cause a
pop-up to show up during review.

This is a fix for:
https://code.google.com/p/ankidroid/issues/detail?id=1575