Remove the .project file from .gitignore.
The .project file was currently ignored and not added to version
control. This causes some problems, including the fact that the newly
added project is called as the first activity named in the manifest,
currently DeckOptions, which is misleading.

Tracking the .project file is actually useful if we want to use some of
the most fancy features, like filtered resources, so it is useful to
track it.