Remove suppressed warning.
We were suppressing the warning here because we were calling a function
available only in API 16. However, that causes crashes on some versions
of Android and using the Compat interface turned out to be a better
solution.

Now, we can remove the suppressed warning since there is no longer a
warning.