Add odid fix to db check
https://github.com/dae/anki/commit/7b62094ea0d99484463b2242cacb6842a43157cd

Added another missing but similar block in the integrity check. 

The addition of this fix allows us to remove a workaround in Sched. When this problem is now triggered, it will bubble up to the general-purpose exception handler in the reviewer which offers the option to do a database check where it can be fixed.