Expose String and Boolean variants of findCards
The _order method already made the type distinction, but findCards is what actually uses that method and it was always dealing with Strings.

Fixes an error where the _order code path was falsely determined on a card search (and thus failing) by using a "true" String instead of a boolean true.