Play sounds with link handler instead of javascript interface
This commit removes the javascript interface.

The previous use of the javascript interface was to handle taps made on the image playback button to play sounds with the device's media player. With this commit, we now do it by handling links added to the playback buttons marked with a prefix specifically for audio playback. In the future, new protocols (or prefixes) can be added to handle any new commands that must be handled in Java code.

The rest of the interface contained some old code for image resizing but is no longer used and can be safely removed.