Add a command-line tool for integrating changes.
The purpose of this tool is to help integrate changes from the
'release-*' and 'hotfix-*' branches back into the 'develop' branch,
while keeping the details of the changes being integrated.

Each change is merged individually into the 'develop' branch, and
changes can be skipped manual or automatically (by adding
"@branch-specific" into the commit message). For skipped changes, an
empty commit is created to record the resolution: this avoids changes
like increment of the version number to lead to conflicts later on.