RecyclerView for DeckPicker
Replaced ListView with a RecyclerView in DeckPicker. This also replaces the anonymous SimpleAdapter with a custom RecyclerView.Adapter. Our UI is a simple list, which does not require any custom styling depending on whether the item is a deck or subdeck. The items have click & long click actions that function the same as before. There is also a new method inside DeckPicker for finding a deck node given a deck ID.