Don't flash white when stats are first opened, change black on black to grey on black in statistics charts
- When you first open the stats, it now doesn't flashes white anymore
for a fraction of a second. Fixed by changing the background to
transparent like is done in the Reviewer.
- When using the black theme, the unseen cards are now shown as grey on
black in the pie chart instead of black
- When using the black theme, the cumulative line in the intervals chart
are now shown as grey on black instead of black on black
- When using the black theme, the counts bars in the hourly and weekly
breakdown chart are now shown as grey on black instead of black on black