Remove latency for long-tap to work in Browser
Previously we were calling TASK_TYPE_RENDER_BROWSER_QA with the first 1000 cards in the result from mSearchCardsHandler.onPostExecute().
Since mRenderQAHandler does a notifyDataSetChanged() for EVERY card in the list as they're updated, this was effectively blocking the long-click listener for the cards, so I moved that initial card rendering code into the first DeckTask and dropped it from 1000 to 500.

I also did some tidying up to use the List and Map interfaces polymorphically instead of the ArrayList and HashMap implementations.

Fixes #4191